requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.8
django-allauth==0.35.0
django-crispy-forms==1.7.0
django-rest-auth==0.9.3
django-taggit==0.22.2
django-widget-tweaks==1.4.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
oauthlib==2.0.6
Pillow==5.0.0
psycopg2==2.7.4
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
whitenoise==3.3.1
